Add edit session type + split create session modal component
