+++ client/src/lib/content/hooks/useExerciseIds.ts
@@ -1,6 +1,7 @@
 import {exerciseIds} from '../../../../../content/content.json';
 import {Exercise} from '../../../../../shared/src/types/generated/Exercise';
 
-const useExerciseIds = () => exerciseIds as Exercise['id'][];
+const useExerciseIds: () => Exercise['id'][] = () =>
+  exerciseIds as Exercise['id'][];
 
 export default useExerciseIds;
