+++ content/src/ui/Modal.CreateSession.json
@@ -20,7 +20,11 @@
     },
     "setDateTime": {
       "title": "Choose a time",
-      "cta": "Create session"
+      "cta": "Create session",
+      "sessionType": {
+        "private": "Session with Friends",
+        "public": "Session with Anyone"
+      }
     },
     "profile": {
       "text": "To create and host a session you need a user name and profile picture."
